extract command_runner
