+++ frb_codegen/src/commands.rs
@@ -1,18 +1,14 @@
-use anyhow::anyhow;
-use itertools::Itertools;
-use log::log_enabled;
 use std::fmt::Write;
 use std::path::Path;
 use std::path::PathBuf;
-use std::process::Command;
-use std::process::Output;
-use std::process::Stdio;
 use std::str::FromStr;
 use lazy_static::lazy_static;
 use cargo_metadata::VersionReq;
 
 use crate::error::{Error, Result};
-use log::{debug, info, warn};
+use log::{debug, info};
+use crate::run;
+use crate::utils::command_runner::{call_shell, execute_command};
 use crate::utils::dart_repository::dart_repository::{DartDependencyMode, DartRepository};
 
 lazy_static! {
@@ -20,78 +16,6 @@ lazy_static! {
         VersionReq::parse(">= 2.0.1, < 3.0.0").unwrap();
     pub(crate) static ref FFIGEN_REQUIREMENT: VersionReq =
         VersionReq::parse(">= 6.0.1, < 8.0.0").unwrap();
-}
-
-/// - First argument is either a string of a command, or a function receiving a slice of [`PathBuf`].
-///   - The command may be followed by `in <expr>` to specify the working directory.
-///   - The function may be followed by an array of rest parameters to pass.
-/// - Following arguments are either:
-///   - An expression to turn into a [`PathBuf`]; or
-///   - `?<expr>` to add `expr` only if `expr` is a [`Some`]; or
-///   - `*<expr>` to concatenate an iterable of such expressions; or
-///   - A tuple of `(condition, expr, ...expr)` that adds `expr`s to the arguments only if `condition` is satisfied.
-///
-/// Returns [`Result<Output>`] if executing a command name, or the return value of the specified function.
-#[doc(hidden)]
-#[macro_export]
-macro_rules! run {
-    ($binary:literal, $($rest:tt)*) => {{
-        let args = $crate::args!($($rest)*);
-        $crate::commands::execute_command($binary, args.iter(), None)
-    }};
-    ($binary:literal in $pwd:expr, $($rest:tt)*) => {{
-        let args = $crate::args!($($rest)*);
-        $crate::commands::execute_command($binary, args.iter(), $pwd)
-    }};
-    ($command:path $([ $($args:expr),* ])?, $($rest:tt)*) => {{
-        let args = $crate::args!($($rest)*);
-        $command(&args[..] $(, $($args),* )?)
-    }};
-}
-
-/// Formats a list of [`PathBuf`]s using the syntax detailed in [`run`].
-#[doc(hidden)]
-#[macro_export]
-macro_rules! args {
-    (@args $args:ident $(,)?) => {};
-    (@args $args:ident ($cond:expr, $($expr:expr),+ $(,)?), $($rest:tt)*) => {
-        if $cond {
-            $(
-                $args.push(::std::path::PathBuf::from($expr));
-            )+
-        }
-        $crate::args!(@args $args $($rest)*);
-    };
-    (@args $args:ident ?$src:expr, $($rest:tt)*) => {
-        if let Some(it) = (&$src) {
-            $args.push(::std::path::PathBuf::from(it));
-        }
-        $crate::args!(@args $args $($rest)*);
-    };
-    (@args $args:ident *$src:expr, $($rest:tt)*) => {
-        $args.extend($src.iter().map(::std::path::PathBuf::from));
-        $crate::args!(@args $args $($rest)*);
-    };
-    (@args $args:ident $expr:expr, $($rest:tt)*) => {
-        $args.push(::std::path::PathBuf::from($expr));
-        $crate::args!(@args $args $($rest)*);
-    };
-    ($($rest:tt)*) => {{
-        let mut args = Vec::new();
-        $crate::args!(@args args $($rest)*,);
-        args
-    }};
-}
-
-pub(crate) fn call_shell(cmd: &[PathBuf], pwd: Option<&str>) -> Result<Output> {
-    let cmd = cmd.iter().map(|section| format!("{section:?}")).join(" ");
-    #[cfg(windows)]
-    {
-        run!("powershell" in pwd, "-noprofile", "-command", format!("& {}", cmd))
-    }
-
-    #[cfg(not(windows))]
-    run!("sh" in pwd, "-c", cmd)
 }
 
 pub fn ensure_tools_available(dart_root: &str, skip_deps_check: bool) -> Result {
@@ -149,70 +73,6 @@ pub(crate) fn bindgen_rust_to_dart(
         arg.llvm_compiler_opts,
         dart_root,
     )
-}
-
-#[must_use = "Error path must be handled."]
-fn execute_command<'a>(
-    bin: &str,
-    args: impl IntoIterator<Item = &'a PathBuf>,
-    current_dir: Option<&str>,
-) -> Result<Output> {
-    let args = args.into_iter().collect::<Vec<_>>();
-    let args_display = args.iter().map(|path| path.to_string_lossy()).join(" ");
-    let mut cmd = Command::new(bin);
-    cmd.args(args);
-    if log_enabled!(log::Level::Info) {
-        cmd.stdout(Stdio::inherit()).stderr(Stdio::inherit());
-    }
-
-    if let Some(current_dir) = current_dir {
-        cmd.current_dir(current_dir);
-    }
-    if log_enabled!(log::Level::Info) {
-        cmd.stdout(Stdio::inherit()).stderr(Stdio::inherit());
-    }
-
-    debug!(
-        "execute command: bin={} args={:?} current_dir={:?} cmd={:?}",
-        bin, args_display, current_dir, cmd
-    );
-
-    let result = cmd
-        .output()
-        .map_err(|err| anyhow!("\"{}\" \"{}\" failed: {}", bin, args_display, err))?;
-
-    let stdout = String::from_utf8_lossy(&result.stdout);
-    if result.status.success() {
-        debug!(
-            "command={:?} stdout={} stderr={}",
-            cmd,
-            stdout,
-            String::from_utf8_lossy(&result.stderr)
-        );
-        if stdout.contains("fatal error") {
-            warn!("See keywords such as `error` in command output. Maybe there is a problem? command={:?} output={:?}", cmd, result);
-        } else if args_display.contains("ffigen") && stdout.contains("[SEVERE]") {
-            // HACK: If ffigen can't find a header file it will generate broken
-            // bindings but still exit successfully. We can detect these broken
-            // bindings by looking for a "[SEVERE]" log message.
-            //
-            // It may emit SEVERE log messages for non-fatal errors though, so
-            // we don't want to error out completely.
-
-            warn!(
-                "The `ffigen` command emitted a SEVERE error. Maybe there is a problem? command={:?} output=\n{}",
-                cmd, String::from_utf8_lossy(&result.stdout)
-            );
-        }
-    } else {
-        warn!(
-            "command={:?} stdout={} stderr={}",
-            cmd,
-            stdout,
-            String::from_utf8_lossy(&result.stderr)
-        );
-    }
-    Ok(result)
 }
 
 fn cbindgen(
