+++ frb_codegen/src/utils/dart_repository/dart_toolchain.rs
@@ -1,5 +1,6 @@
-use crate::commands::call_shell;
+use crate::utils::command_runner::call_shell;
 use std::path::PathBuf;
+use crate::{args, run};
 
 /// represents dart or flutter toolchain
 #[derive(Debug, PartialEq, Eq)]
