+++ frb_codegen/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod consts;
 pub(crate) mod misc;
 pub(crate) mod dart_repository;
+pub(crate) mod command_runner;
